feat: Implement and integrate an affiliate rebate system, applying rewards after successful top-ups.

+++ controller/midjourney.go
@@ -181,7 +181,7 @@ func UpdateMidjourneyTaskBulk() {
 						if err != nil {
 							logger.LogError(ctx, "fail to increase user quota: "+err.Error())
 						}
-						logContent := fmt.Sprintf("构图失败 %s，补偿 %s", task.MjId, logger.LogQuota(task.Quota))
+						logContent := fmt.Sprintf("构图失败 %s，补偿 %s", task.MjId, logger.LogQuota(int64(task.Quota)))
 						model.RecordLog(task.UserId, model.LogTypeSystem, logContent)
 					}
 				}
